+++ ceammc/CHANGELOG.md
@@ -14,6 +14,7 @@
 - random.discrete @seed property added to gen reproducible random sequences
 - ui.toggle @on_value/@off_value properties added
 - @float_width property added to ui.display
+- @positive property added to car->pol for output in [0..2π) range
 
 ### Changed:
 - ext_info output format changed.